chore: move DatabaseUpdateMaintenanceWindow to maintenance.go
