+++ databases.go
@@ -231,22 +231,3 @@ func (c *Client) DatabaseDisableFeature(ctx context.Context, app, addonID, featu
 
 	return res, nil
 }
-
-type MaintenanceWindowParams struct {
-	WeekdayUTC      *int `json:"weekday_utc,omitempty"`
-	StartingHourUTC *int `json:"starting_hour_utc,omitempty"`
-}
-
-func (c *Client) DatabaseUpdateMaintenanceWindow(ctx context.Context, app, addonID string, params MaintenanceWindowParams) (Database, error) {
-	var dbRes DatabaseRes
-	err := c.DBAPI(app, addonID).ResourceUpdate(ctx, "databases", addonID, map[string]interface{}{
-		"database": map[string]interface{}{
-			"maintenance_window": params,
-		},
-	}, &dbRes)
-
-	if err != nil {
-		return Database{}, errgo.Notef(err, "update database '%v' maintenance window", addonID)
-	}
-	return dbRes.Database, nil
-}
